+++ cli/src/flows/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Flows module exports.
+ */
+
+export { flowEtsiTrustLists } from './etsi.js';
+export { flowCredentialRevocation } from './revocation.js';
