+++ cli/src/http/index.ts
@@ -1,0 +1,5 @@
+/**
+ * HTTP module exports.
+ */
+
+export { HttpClient, type HttpResponse, type HttpLogEntry } from './client.js';
