+++ cli/tsconfig.json
@@ -15,6 +15,6 @@
     "declarationMap": true,
     "sourceMap": true
   },
-  "include": ["*.ts"],
-  "exclude": ["node_modules", "dist", "journey-test-*"]
+  "include": ["*.ts", "src/**/*.ts"],
+  "exclude": ["node_modules", "dist", "logs"]
 }
